Create Chainlink-Project
This is Project is about using chainlink oracles to get real-time value of the ether in terms of US dollar. so user will predict whether price of ether will cross $235 mark in one month or not . for making prediction users  have to buy the shares, if they market resolve to yes they will be paid out the money.